feat: production readiness - security hardening, SEO, legal compliance, UX
Security:
- Add EnsureAdmin middleware for admin route guards
- Fix cross-tenant policy bugs in AIRecommendation/ShiftTemplate policies
- Add HasTeamFilters null-team bypass fix
- Add EnsureValidRealTimeChannel middleware for broadcast auth
- Add CacheControlHeaders middleware (private cache for auth responses)
- Add ForceHttps middleware (301 redirect in production)
- Fix Stripe webhook signature validation in WebhookController
- Add enterprise security indexes migration
- Add channel authorization in routes/channels.php
- Restrict test routes to local+debug only

Bug fixes (static analysis):
- Fix preg_replace nullable returns, strtoupper type coercion
- Fix UploadedFile null guard in FeedAttachmentService
- Fix Machine.activeEngineSession latestOfMany chain
- Fix AuditService wrong argument order
- Fix fopen unchecked return in GenerateReportJob and Reports.php
- Fix strlen on nullable, Auth::id() int cast, redirect() string cast
- Fix model docblock positions (MaintenanceRecord, FeedComment, FeedPost,
  FeedAttachment, ShiftTemplate, ProductionRecord, Subscription)
- Regenerate PHPStan baseline to 0 errors

Reports:
- Fix route ordering in web.php and api.php
- Fix missing Storage import in Api\ReportController
- Fix show.blade.php download URL using temporary signed routes
- Fix ReportController::view2() cross-tenant fallback
- Add GenerateReportJob with PDF (dompdf) and CSV generation

Legal / compliance:
- Enable Jetstream termsAndPrivacyPolicy feature (privacy-policy + terms-of-service routes)
- Add cookie consent banner component (Alpine.js + localStorage, POPIA/GDPR)
- Verify account deletion (right to erasure) via Jetstream accountDeletion

UX:
- Add custom error pages: 404, 403, 500, 419
- Add copyright footer with Privacy Policy and Terms of Service links

SEO:
- Add SEO meta tags to all layouts (description, canonical, Open Graph, Twitter Card)
- noindex/nofollow on authenticated app layouts
- Add sitemap.xml route with all public pages
- Update robots.txt: disallow all app/admin/api routes, add Sitemap directive

Performance:
- Add deploy/nginx.conf with Cache-Control immutable for Vite /build/ assets,
  GZIP compression, TLS 1.2/1.3, static asset expiry headers

Email:
- Configure SMTP for info@mines.infodot.co.za (SSL port 465)
- Update .env.example and .env.production.example

+++ app/Livewire/Feed.php
@@ -342,7 +342,7 @@ class Feed extends Component
         $post->refresh();
 
         // Parse @mentions in comment
-        app(MentionParser::class)->parseSave($comment, $comment->body, Auth::id(), $post->team_id);
+        app(MentionParser::class)->parseSave($comment, $comment->body, (int) Auth::id(), (int) $post->team_id);
 
         FeedCommentCreated::dispatch($comment, $post);
